+++ tests/test_spec.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import tempfile
 import unittest
+from unittest.mock import Mock, patch
 
 import geometamaker
 import numpy
@@ -335,8 +336,8 @@ class ResultsSuffixTests(unittest.TestCase):
         self.assertEqual(missing_suffix, [])
 
 
-class MissingResultsSuffixTests(unittest.TestCase):
-    """Test ModelSpec.execute for model without ResultsSuffixInput."""
+class ModelSpecExecuteTests(unittest.TestCase):
+    """Test ModelSpec.execute."""
 
     def setUp(self):
         """Override setUp function to create temp workspace directory."""
@@ -395,6 +396,48 @@ class MissingResultsSuffixTests(unittest.TestCase):
         carbon.MODEL_SPEC.execute(args, **execute_kwargs)
         assert_complete_execute(
             args, carbon.MODEL_SPEC, **execute_kwargs)
+
+    @patch('natcap.invest.carbon.carbon.execute')
+    def test_execute_function_returns_none(self, mock_execute):
+        """ModelSpec.execute tolerates missing file_registry dict."""
+        from natcap.invest import carbon
+        mock_execute = Mock()
+        mock_execute.return_value = None
+
+        execute_kwargs = {
+            'check_outputs': True,
+            'generate_metadata': True,
+            'save_file_registry': True,
+            'generate_report': True
+        }
+        kwargs_used = [f'{k}={v}' for k, v in execute_kwargs.items()]
+        message = "`execute` was called with {kwargs}, but no" \
+                  " file registry dictionary was returned from `execute`."
+
+        with self.assertLogs('natcap.invest.spec', level='WARNING') as cm:
+            carbon.MODEL_SPEC.execute({}, **execute_kwargs)
+        self.assertIn(
+            message.format(kwargs=', '.join(kwargs_used)),
+            ' '.join(cm.output))
+
+    @patch('natcap.invest.carbon.carbon.execute')
+    def test_execute_function_returns_empty_dict(self, mock_execute):
+        """ModelSpec.execute tolerates an empty file_registry dict."""
+        from natcap.invest import carbon
+        mock_execute = Mock()
+        mock_execute.return_value = {}
+
+        execute_kwargs = {
+            'check_outputs': True,
+            'generate_metadata': True,
+            'save_file_registry': True,
+            'generate_report': True
+        }
+
+        try:
+            carbon.MODEL_SPEC.execute({}, **execute_kwargs)
+        except Exception as error:
+            self.fail(error)
 
 
 class InputTests(unittest.TestCase):